#460: moved element testing to it's own file in the tests dir.
